Create BiometricsConfigRepository to sync
