+++ app/src/main/assets/paperwork.json
@@ -1,1 +1,1 @@
-{"buildTime":"2021-08-06 06:35","gitSha":"b002530d0"}
+{"buildTime":"2021-08-06 08:56","gitSha":"ad3936686"}
