+++ app/src/main/java/org/dhis2/usescases/biometrics/BiometricsConfigRepository.kt
@@ -1,4 +1,5 @@
 package org.dhis2.usescases.biometrics
 
 interface BiometricsConfigRepository {
+    fun sync()
 }
